Organize classes in the appropriate modules.
