+++ qubecalib/qubecalib/backendqube.py
@@ -2,17 +2,20 @@ from .qube import CPT, AWG, UNIT
 from .meas import WaveSequenceFactory, CaptureModule, CaptureParam
 from .setupqube import _conv_to_e7awgsw, _conv_channel_for_e7awgsw
 from .pulse import Read, Arbit
-from .neopulse import WORDs, nS, MHz
-import e7awgsw
-from e7awgsw import WaveSequence
-from typing import Union
+from .neopulse import Series, Blank, Range, Capture, ContextNode, Readout, body, SlotWithIQ
+from .units import *
 
+import e7awgsw
+from e7awgsw import WaveSequence, CaptureParam, AwgCtrl, IqWave
+from typing import Union
 import os
 import time
 import struct
-from collections import namedtuple
+from collections import namedtuple, deque
 from concurrent.futures import ThreadPoolExecutor
-from contextlib import redirect_stdout
+from contextlib import redirect_stdout, contextmanager
+import matplotlib.pyplot as plt
+import matplotlib.patches as patches
 import warnings
 import datetime
 import numpy as np
@@ -31,6 +34,7 @@ IPADDR = '10.3.0.255'
 REPEAT_WAIT_SEC = 0.1
 REPEAT_WAIT_SEC = int(REPEAT_WAIT_SEC * 125000000) # 125Mcycles = 1sec
 CANCEL_STOP_PACKET = struct.pack(8*'B', 0x2c, *(7*[0]))
+
 
 
 def check_clock(*qubes, ipmaster='10.3.0.255'):
@@ -892,3 +896,695 @@ def captparam_enable_demodulation(captparam, physical_channel, logical_channel):
     # dspunits.append(e7awgsw.DspUnit.REAL_FIR)
     dspunits.append(e7awgsw.DspUnit.COMPLEX_WINDOW) # 複素窓関数を有効化
     p.sel_dsp_units_to_enable(*dspunits)
+
+class Setup(list):
+
+    class CaptParamsIter( object ):
+
+        def __init__(self, setup, channel_map, *logical_channels):
+
+            self.setup = setup
+            self.channel_map = channel_map
+            if logical_channels:
+                self.logical_channels = logical_channels
+            else:
+                self.logical_channels = tuple([channel_map.logical[u][0] for u, c in setup if isinstance(u, UNIT) and u in channel_map.logical])
+            self.current_idx = 0
+        
+        def __iter__(self):
+
+            return self
+
+        def __next__(self):
+
+            # while True:
+
+            #     try:
+            #         logical_channel = l = self.logical_channels[self.current_idx]
+            #     except IndexError:
+            #         raise StopIteration()
+
+            #     self.current_idx += 1
+
+            #     try:
+            #         self.channel_map.physical[l]
+            #     except KeyError:
+            #         pass
+            #     else:
+            #         break
+                    
+            try:
+                logical_channel = l = self.logical_channels[self.current_idx]
+            except IndexError:
+                raise StopIteration()
+
+            u = [o for o in self.channel_map.physical[l] if isinstance(o, UNIT)][0]
+
+            self.current_idx += 1
+
+            return self.setup.get(u), u, l
+
+    def __init__(self,*args):
+
+        super().__init__(args)
+
+    def get(self,arg):
+
+        keys = [k for k,v in self]
+        return self[keys.index(arg)][1]
+
+    def captparams(self, channel_map, *logical_channels):
+
+        return Setup.CaptParamsIter(self, channel_map, *logical_channels)
+
+class ContextNodeAlloc(object):
+
+    def __init__(
+        self,
+        begin,
+        duration=None,
+        end=None,
+    ):
+        
+        self.begin = begin
+        self._end_duration(end,duration)
+        
+    def _end_duration(self,end,duration):
+        if duration is None:
+            self.end = end
+            self.duration = end - self.begin
+        elif end is None:
+            self.end = self.begin + duration
+            self.duration = duration
+        else:
+            raise()
+        
+    def alloc(self,*args):
+
+        c = __rc__.contexts
+        if len(c):
+            d = c[-1]
+            for a in args:
+                if a not in d:
+                    d[a] = deque([self.__class__(self.begin,self.duration)])
+                else:
+                    d[a].append(self.__class__(self.begin,self.duration))
+
+
+class Chunk(ContextNodeAlloc):
+
+    def __init__(
+        self,
+        begin,
+        duration=None,
+        end=None,
+    ):
+
+        super().__init__(begin,duration,end)
+        duration_in_samples = n = int(self.duration // 2)
+        self.iq = np.zeros(n).astype(complex)
+    
+    @property
+    def sampling_points(self):
+        
+        return np.arange(len(self.iq)) * 2 + self.begin # sampling points [ns]
+
+class SumSect(ContextNodeAlloc): pass
+
+
+def acquire_section(sequence, channel_map, section=None):
+
+    if section is None:
+        section = Sections()
+
+    tx = acquire_tx_section(sequence, channel_map)
+    rx = acquire_rx_section(sequence.flatten(), channel_map)
+
+    for k, v in tx.items():
+        section[k] = v
+
+    for k, v in rx.items():
+        section[k] = v
+
+    section.place()
+
+    return section
+
+def acquire_tx_section(sequence, channel_map, section=None):
+
+    if section is None:
+        section = Sections()
+
+    logch = [k for k in sequence.flatten().slots if k is not None]
+    phych = [k for k in channel_map for l in logch if l in channel_map[k] and isinstance(k,AWG)]
+
+    is_multi_chunk = np.array([isinstance(o, Series) for o in sequence]).prod()
+    is_multi_chunk *= np.array([isinstance(v, (Series, Blank)) for o in sequence if isinstance(o, Series) for v in o]).prod()
+
+    if is_multi_chunk:
+
+        for phychi in phych:
+
+            section[phychi] = deque()
+
+            wait = 0
+            series = [v for o in sequence if isinstance(o, Series) for v in o]
+            for o in [v for o in sequence if isinstance(o, (Series, Blank)) for v in o]:
+
+                if isinstance(o, Blank):
+
+                    wait += o.duration
+                    
+                elif isinstance(o, Series):
+                    
+                    d = (o.end - o.begin) / o.repeats
+
+                    if o is not series[-1]:
+
+                        blank = d % BLOCK
+                        if blank:
+                            u = [v for v in o.bodies.flatten() if not isinstance(v, (Range, Blank))]
+                            b = min([v.begin for v in u])
+                            e = max([v.end for v in u])
+                            if o.end / o.repeats - e < blank:
+                                raise ValueError('duration is wrong')
+                        d = d // BLOCK * BLOCK
+                        section[phychi].append(TxSection(wait=wait, duration=d, blank=blank, repeats=o.repeats))
+
+                    else:
+                        
+                        d = d // BLOCK * BLOCK + BLOCK
+                        section[phychi].append(TxSection(wait=wait, duration=d, blank=0, repeats=o.repeats))
+                    
+                    wait = 0
+
+    else:
+        
+        # print('Single Chunk Mode')
+        for phychi in phych:
+
+            section[phychi] = deque()
+
+            u = [v for v in sequence.flatten() if not isinstance(v, (Range, Blank))]
+            b = min([v.begin for v in u]) // WORD * WORD
+            e = max([v.end for v in u])
+            d = ((e - b) // BLOCK + 1) * BLOCK
+            section[phychi].append(TxSection(wait=b, duration=d, blank=0, repeats=1))
+    
+    return section
+
+
+    # # log2phy = {o: k for k,v in channel_map.items() if isinstance(k,AWG) for o in v}
+    # log2phy = {k: [o for o in v if isinstance(o,AWG)][0] for k, v in channel_map.logical.items()}
+    # logch = [k for k in sequence.flatten().slots if k is not None]
+    # phych = [k for k in channel_map for l in logch if l in channel_map[k] and isinstance(k,AWG)]
+
+    # # AWG と Series の対応辞書を作る
+    # series = find_sequence(sequence, Series)
+    # awg2series = d = dict()
+    # for p in phych:
+    #     d[p] = deque()
+    #     for s in series:
+    #         if p in [log2phy[k] for k in s.flatten().slots.keys() if k is not None]:
+    #             d[p].append(s)
+
+    # # 重複した Series を削除する
+    # for k, v in awg2series.items():
+    #     item = lambda x: (x.begin, (x.end - x.begin) / x.repeats, x)
+    #     rng = r = sorted([item(o) for o in v], key=lambda x:x[1], reverse=True)
+    #     buffer = b = Sequence()
+    #     isout = lambda x, y: (y[0] + y[1] < x[0]) + (x[0] + x[1] < y[0])
+    #     isin = lambda x, y: (x[0] < y[0]) * (y[0] + y[1] < x[0] + x[1])
+    #     for i, o in enumerate(rng[:-1]):
+    #         skip = False
+    #         for m in rng[i+1:]:
+    #             if skip:
+    #                 continue
+    #             if isin(o, m):
+    #                 skip = True
+    #                 continue
+    #             b.append(o[2])
+    #     awg2series[k] = buffer
+        
+    # # AWG と Slots の対応辞書を作る
+    # awg2slots = dict()
+    # slots = sequence.flatten().slots
+    # for phychi in phych:
+    #     for k, v in slots.items():
+    #         if phychi is log2phy[k] if k is not None else None:
+    #             awg2slots[phychi] = deque()
+    #             for o in v:
+    #                 if isinstance(body(o),SlotWithIQ):
+    #                     if o not in awg2series[phychi].flatten():
+    #                         awg2slots[phychi].append(o)
+
+    # 最も浅い repeats 付きの Series を参考にして Chunk をつくる
+    
+
+    # if section is None:
+    #     section = Sections()
+
+    # for phychi in phych:
+
+    #     print(phychi, awg2series[phychi])
+    #     slots = np.array([(s.begin,(s.end-s.begin)/s.repeats,s.end,s.repeats) for s in awg2series[phychi] if isinstance(s,Series)])
+    #     print(phychi, slots)
+    #     if not slots:
+    #         if not awg2slots[phychi]:
+    #             continue
+    #         begin = min([o.begin for o in awg2slots[phychi]])
+    #         end = max([o.end for o in awg2slots[phychi]])
+    #         if begin % int(WORD):
+    #             print(phychi, begin)
+    #             raise ValueError('begin is not aligned')
+    #         if (end - begin) % int(BLOCK):
+    #             raise ValueError('durations is not aligned')
+    #         section[phychi] = deque()
+    #         section[phychi].append(TxSection(wait=begin,duration=end-begin,blank=0,repeats=1))
+    #         continue
+    #     slots = slots[np.argsort(slots[:,0])]
+
+    #     section[phychi] = deque()
+
+    #     begin, duration, end, repeats = slots[0]
+    #     section[phychi].append(TxSection(wait=begin,duration=duration,blank=0,repeats=int(repeats)))
+    #     for i in range(1,slots.shape[0]):
+    #         begin, duration, end, repeats = slots[i,0], slots[i,1], slots[i-1,2], slots[i,3]
+    #         section[phychi].append(TxSection(wait=begin-end,duration=duration,blank=0,repeats=int(repeats)))
+
+    # return section
+
+
+def acquire_rx_section(sequence, channel_map, section=None):
+
+    if section is None:
+        section = Sections()
+
+    logch = [k for k in sequence.slots if isinstance(k,Readout)]
+    phych = [k for k in channel_map for l in logch if l in channel_map[k] and isinstance(k,UNIT)]
+
+    for phychi in phych:
+
+        slots = np.array([(s.begin,s.duration,s.end) for s in sequence if isinstance(body(s),Range) and s.ch in channel_map[phychi]])
+        slots = slots[np.argsort(slots[:,0])]
+        
+        # duration % WORD == 0, blank % WORD == 0
+        # begin が条件を満たすように Range を前に offset する
+        # duration が条件を満たすように Range を stretch する
+        begin, duration, end = slots[0]
+        offset = begin - (begin // (2*WORD) * (2*WORD))
+        begin = begin - offset
+        duration = offset + duration
+        stretch = duration - (duration // WORD * WORD)
+        duration = duration + stretch
+        end = end + stretch
+        slots[0,:] = [begin, duration, end]
+        for i, slot in enumerate(slots[1:]):
+            begin, duration, end = slot
+            offset = begin - (begin // WORD * WORD)
+            begin = begin - offset
+            duration = offset + duration
+            stretch = duration - (duration // WORD * WORD)
+            duration = duration + stretch
+            end = end + stretch
+            slots[i+1,:] = [begin, duration, end]
+            
+        for i, _ in enumerate(slots[:-1]):
+            b, d, e = slots[i+1][0], slots[i+1][1], slots[i][2]
+            if b - e < 0:
+                raise ValueError('Sumsections are overlapped.')
+        section[phychi] = deque()
+
+        begin, duration, end = slots[0]
+        section[phychi].append(RxSection(delay=begin,duration=duration,blank=0))
+        for i in range(1,slots.shape[0]):
+            begin, duration, end = slots[i,0], slots[i,1], slots[i-1,2]
+            section[phychi].append(RxSection(delay=begin-end,duration=duration,blank=0))
+            
+    return section
+
+class DictWithContext(dict):
+
+    def __enter__(self):
+
+        __rc__.contexts.append(deque())
+
+        return self
+
+    def __exit__(self, exception_type, exception_value, traceback):
+
+        __rc__.contexts.pop()
+
+
+class Sections(DictWithContext):
+    
+    def __init__(self,*args,**kw):
+        
+        super().__init__(**kw)
+        for k in args:
+            self[k] = None
+
+    def __exit__(self, exception_type, exception_value, traceback):
+
+        q = __rc__.contexts.pop()
+        for k in self:
+            if isinstance(q,Sections):
+                raise ValueError('Deep nesting is not allowed.')
+            self[k] = q
+        if len(self):
+            __rc__.contexts[-1].append(self)
+        else:
+            for o in q:
+                for k,v in o.items():
+                    if k in self:
+                        raise KeyError('Key dupulication is detected.')
+                    self[k] = v
+        self.place()
+
+    def place(self):
+        
+        for k, v in self.items():
+            if v:
+                v0 = v[0]
+                v0.begin = v0.prior
+                v0.end = v0.begin + v0.duration
+                for i in range(1,len(v)):
+                    total = v[i-1].prior + v[i-1].duration + v[i-1].post
+                    v[i].begin = v[i-1].end + v[i-1].post + (v[i-1].repeats - 1) * total + v[i].prior
+                    v[i].end = v[i].begin + v[i].duration
+
+
+class Allocation(DictWithContext):
+    pass
+
+
+class AllocTable(dict): pass
+
+
+def plot_send_recv(fig, data, mag=False):
+
+    n = len([vv for k,v in data.items() for vv in v])
+    i = 1
+    rslt = []
+    for k,v in data.items():
+        for vv in v:
+            if fig.axes:
+                ax = fig.add_subplot(n,1,i, sharex=ax1)
+            else:
+                ax = ax1 = fig.add_subplot(n,1,i)
+            if mag:
+                ax.plot(np.abs(vv))
+            else:
+                ax.plot(np.real(vv))
+                ax.plot(np.imag(vv))
+            i += 1
+            rslt.append(k)
+    return rslt
+
+def plot_setup(fig,setup,capture_delay=0):
+    for i,tpl in enumerate(setup):
+        k,v = tpl
+        if i == 0:
+            ax = ax1 = fig.add_subplot(len(setup),1,i+1)
+        else:
+            ax = fig.add_subplot(len(setup),1,i+1, sharex=ax1)
+        if isinstance(k,AWG):
+            blank = [o.num_blank_words * int(WORDs) for o in v.chunk_list]
+            begin = v.num_wait_words * int(WORDs)
+            for cc,bb in zip(v.chunk_list,blank):
+                iq = np.array(cc.wave_data.samples)
+                t = begin + np.arange(len(iq)) * 2
+                for _ in range(cc.num_repeats-1):
+                    ax.plot(t,iq[:,0],'b')
+                    ax.plot(t,iq[:,1],'r')
+                    ax.set_ylim(-32767*1.2,32767*1.2)
+                    t += len(iq) * 2 + bb
+                ax.plot(t,iq[:,0],'b')
+                ax.plot(t,iq[:,1],'r')
+                ax.set_ylim(-32767*1.2,32767*1.2)
+                begin = t[-1] + bb
+        else:
+            blank = [o[1] * int(WORDs) for o in v.sum_section_list]
+            begin = (v.capture_delay - capture_delay) * int(WORDs)
+            for s,b in zip(v.sum_section_list,blank):
+                duration = int(s[0]) * int(WORDs)
+                ax.add_patch(patches.Rectangle(xy=(begin,-32767),width=duration,height=2*32767))
+                begin += duration + b
+                ax.set_ylim(-32767*1.2,32767*1.2)
+
+def plot_section(fig,section):
+
+    for i,phych in enumerate(section):
+        if i == 0:
+            ax = ax1 = fig.add_subplot(len(section),1,i+1)
+        else:
+            ax = fig.add_subplot(len(section),1,i+1,sharex=ax1)
+        for s in section[phych]:
+            begin = int(s.begin)
+            duration = int(s.duration)
+            total = int(s.prior + s.duration + s.post)
+            for i in range(s.repeats):
+                ax.add_patch(patches.Rectangle(xy=(begin + i * total,-1),width=duration,height=2,fill=False))
+        ax.set_xlim(0,s.end)
+        ax.set_ylim(-1.2,1.2)
+
+
+@contextmanager
+def new_section():
+
+    __rc__.contexts.append(Sections())
+    try:
+        yield __rc__.contexts[-1]
+    finally:
+        __rc__.contexts.pop()
+
+
+class SectionBase( ContextNode ):
+
+    def __init__(self, prior=0, duration=BLOCK, post=0, repeats=1):
+        
+        super().__init__()
+        self.prior = prior
+        self.duration = duration
+        self.post = post
+        self.repeats = repeats
+
+    @property
+    def total(self):
+
+        return self.prior + self.duration + self.post
+
+
+class TxSection( SectionBase ):
+
+    def __init__(self, wait=0, duration=BLOCK, blank=0, repeats=1):
+
+        super().__init__(wait, duration, blank, repeats)
+        self._iq = None
+
+    @property
+    def iq(self):
+
+        if self._iq is None:
+            n = int(self.duration // 2)
+            self._iq = np.zeros(n).astype(complex)
+        return self._iq
+    
+    @property
+    def sampling_points(self):
+
+        return np.arange(len(self.iq)) * 2 + self.begin
+
+class RxSection( SectionBase ):
+
+    def __init__(self, delay=0, duration=BLOCK, blank=0, repeats=1):
+
+        super().__init__(delay, duration, blank, repeats)
+
+
+def split_awg_unit(channel_map):
+    a,c = {},{}
+    for k,v in channel_map.items():
+        for vv in v:
+            if isinstance(k,UNIT):
+                c[vv] = k
+            else:
+                a[vv] = k
+    return a,c
+
+def __acquire_chunk__(chunk, repeats=1, blank=0):
+
+    n = WaveSequence.NUM_SAMPLES_IN_WAVE_BLOCK
+    i, q = np.real(32767*chunk.iq).astype(int), np.imag(32767*chunk.iq).astype(int)
+    s = IqWave.convert_to_iq_format(i, q, n)
+
+    b = max(int(blank), 0)
+
+    return {
+        'iq_samples': s,
+        'num_blank_words': b // int(WORD),
+        'num_repeats': repeats
+    }
+
+def chunks2wseq(chunks, period, repeats):
+    # c = chunks
+    # if int(c[0].begin) % WORD:
+    #     raise('The wait duration must be an integer multiple of ()ns.'.format(int(WORD)))
+    # w = e7awgsw.WaveSequence(num_wait_words=int(c[0].begin) // int(WORD), num_repeats=repeats)
+    # boundary = b = [int(cc.begin) - int(c[0].begin) for cc in list(c)[1:]] + [period,]
+    # for s,p in zip(c,b):
+    #     if (p - (s.end - c[0].begin)) % WORD:
+    #         raise('The blank duration must be an integer multiple of ()ns.'.format(int(WORD)))
+    #     kw = {
+    #         'iq_samples': list(zip(np.real(32767*s.iq).astype(int), np.imag(32767*s.iq).astype(int))),
+    #         'num_blank_words': int(p - (s.end - c[0].begin)) // int(WORD),
+    #         'num_repeats': 1,
+    #     }
+    #     w.add_chunk(**kw)
+
+    c = chunks
+    w = e7awgsw.WaveSequence(num_wait_words=int(c[0].prior) // int(WORD), num_repeats=repeats)
+
+    # print(c)
+
+    if len(c) > 1:
+        # ここは nS 単位で． __acquire_chunk__ 内で // int(WORD) してる
+        for j, d in enumerate(list(c)[:-1]):
+            if d.repeats > 1:
+                w.add_chunk(**__acquire_chunk__(d, d.repeats-1, d.post + d.prior))
+                # print((d.post + d.prior), (d.post + d.prior) // int(WORD))
+            w.add_chunk(**__acquire_chunk__(d, 1, d.post + c[j+1].prior))
+            # print((d.post + c[j+1].prior), (d.post + c[j+1].prior) // int(WORD))
+        d = c[-1]
+        if d.repeats > 1:
+            w.add_chunk(**__acquire_chunk__(d, d.repeats-1, d.post + d.prior))
+            # print((d.post + d.prior), (d.post + d.prior) // int(WORD))
+        w.add_chunk(**__acquire_chunk__(d, 1, period - d.end + c[0].prior))
+        # print((period - d.end), (period - d.end) // int(WORD))
+    else:
+        d = c[0]
+        # print('LEN=1', d.prior, d.duration, d.end)
+        w.add_chunk(**__acquire_chunk__(d, 1, period - d.end + c[0].prior))
+    
+    return w
+
+
+def sect2capt(section,period,repeats):
+    s = section
+    p = e7awgsw.CaptureParam()
+    p.num_integ_sections = repeats
+
+    if int(s[0].prior) % int(2*WORD):
+        raise('The capture_delay must be an integer multiple of ()ns.'.format(int(2*WORD)))
+    p.capture_delay = int(s[0].prior) // int(WORD)
+
+    for j in range(len(s)-1):
+        # for i in range(s[j].repeats-1):
+        #     duration = int(s[j].duration) // int(WORD)
+        #     blank = int(s[j].post + s[j].prior) // int(WORD)
+        #     p.add_sum_section(num_words=duration, num_post_blank_words=blank)
+        duration = int(s[j].duration) // int(WORD)
+        blank = int(s[j].post + s[j+1].prior) // int(WORD)
+        p.add_sum_section(num_words=duration, num_post_blank_words=blank)
+    # for i in range(s[-1].repeats-1):
+    #     duration = int(s[-1].duration) // int(WORD)
+    #     blank = int(s[-1].post + s[-1].prior) // int(WORD)
+    #     p.add_sum_section(num_words=duration, num_post_blank_words=blank)
+    duration = int(s[-1].duration) // int(WORD)
+    blank = int(period - s[-1].end + s[0].prior) // int(WORD)
+    p.add_sum_section(num_words=duration, num_post_blank_words=blank)
+
+    return p
+
+
+def convert(sequence, section, channel_map, period, repeats=1, warn=False):
+
+    # channel2slot = r = organize_slots(sequence) # channel -> slot
+    channel2slot = r = sequence.slots
+    a, c = split_awg_unit(channel_map) # channel -> txport, rxport
+    
+    # Tx チャネルはデータ変調
+    for k in a:
+        if k not in r:
+            with warnings.catch_warnings():
+                if not warn:
+                    warnings.simplefilter('ignore')
+                warnings.warn('Channel {} is Iqnored.'.format(k))
+            continue
+        for v in r[k]:
+            if isinstance(body(v),SlotWithIQ):
+                m = a[k].modulation_frequency(mhz=k.frequency*1e+3)*MHz
+                t = v.sampling_points
+                v.miq = v.iq * np.exp(1j*2*np.pi*(m*t))
+            # 位相合わせについては後日
+
+    # 各AWG/UNITの各セクション毎に属する Slot のリストの対応表を作る
+    section2slots = {}
+    for k, v in section.items(): # k:AWG,v:List[TxSection] or k:UNIT,v:List[RxSection]
+        for vv in v: # vv:Union[TxSection,RxSection]
+            if vv not in section2slots:
+                section2slots[vv] = deque([])
+            if k not in channel_map: # k:Union[AWG,UNIT] 
+                with warnings.catch_warnings():
+                    if not warn:
+                        warnings.simplefilter('ignore')
+                    warnings.warn('AWG|UNIT {} is Iqnored.'.format(k))
+                continue
+            for c in channel_map[k]: # c:Channel
+                if c not in r: # sequence の定義内で使っていない論理チャネルがあり得る
+                    with warnings.catch_warnings():
+                        if not warn:
+                            warnings.simplefilter('ignore')
+                        warnings.warn('Channel {} is Iqnored.'.format(c))
+                    continue
+                # for s in r[c]:
+                #     for i in range(vv.repeats):
+                #         print(vv, s, vv.begin + i * vv.total, s.begin, s.begin + s.duration, vv.end + i * vv.total)
+                #         if vv.begin + i * vv.total <= s.begin and s.begin + s.duration <= vv.end + i * vv.total:
+                #             bawg = isinstance(k,AWG) and isinstance(s,SlotWithIQ)
+                #             bunit = isinstance(k,UNIT) and isinstance(body(s),Range)
+                #             if bawg:
+                #                 section2slots[vv].append(s)
+                #             if bunit:
+                #                 section2slots[vv].append(s)
+                for s in r[c]:
+                    if vv.repeats == 1:
+                        bawg = isinstance(k,AWG) and isinstance(body(s),SlotWithIQ)
+                    else:
+                        bawg = isinstance(k,AWG) and isinstance(s,SlotWithIQ)
+                    bunit = isinstance(k,UNIT) and isinstance(body(s),Range)
+                    if bawg:
+                        # print(vv, s, vv.begin, s.begin, s.begin + s.duration, vv.end)
+                        if vv.begin <= s.begin and s.begin <= vv.end:
+                            section2slots[vv].append(s)
+                    elif bunit:
+                        for i in range(vv.repeats):
+                            # print(vv, s, vv.begin + i * vv.total, s.begin, s.begin + s.duration, vv.end + i * vv.total)
+                            if vv.begin + i * vv.total <= s.begin and s.begin + s.duration <= vv.end + i * vv.total:
+                                section2slots[vv].append(s)
+
+
+
+    # for k,v in section2slots.items():
+    #     print(k,v)
+
+    # 各セクション毎に Chunk を合成する
+    awgs = [k for k in channel_map if isinstance(k,AWG)]
+    for i, k in enumerate(awgs):
+        if k in section:
+            for s in section[k]:
+                t = s.sampling_points
+                s.iq[:] = 0
+                ss = section2slots[s]
+                for v in ss:
+                    rng = (v.begin <= t) * (t < v.end)
+                    s.iq[rng] += v.miq / len(ss)
+    
+    # # 束ねるチャネルを WaveSequence に変換
+    awgs = [k for k in channel_map if isinstance(k,AWG)] # channel_map から AWG 関連だけ抜き出す
+    wseqs = [(k, chunks2wseq(section[k], period, repeats)) for k in awgs if k in section] # chunk obj を wseq へ変換する
+
+    units = [k for k in channel_map if isinstance(k,UNIT)] 
+    capts = [(k,sect2capt(section[k],period,repeats)) for k in units if k in section]
+    
+    return Setup(*(wseqs + capts))
